Share Codex approval bridge with tests
